Moved types into their own folder
Without this, can't bring navigation.d.ts in line with new Definitely
Typed spec because importing 'navigation' picks up Navigation.ts instead
of navigation.d.ts
