+++ NavigationReact/test/navigation-react-tests.tsx
@@ -1,7 +1,7 @@
 /// <reference path="../src/navigation.d.ts" />
 /// <reference path="../src/react.d.ts" />
 import { StateNavigator } from 'navigation';
-import { NavigationBackLink, NavigationLink, RefreshLink } from '../src/navigation-react';
+import { NavigationBackLink, NavigationLink, RefreshLink } from '../types/navigation-react';
 import * as React from 'react';
 
 namespace NavigationReactTests {
